+++ src/App.tsx
@@ -111,6 +111,7 @@ import { Toolbar } from './ui/Toolbar'
 import { WalkHud } from './ui/WalkHud'
 import { WallAccentPicker } from './ui/WallAccentPicker'
 import { WebGLFallback } from './ui/WebGLFallback'
+import { WalkCameraControls } from './ui/walk/WalkCameraControls'
 import { WalkJoystick } from './ui/walk/WalkJoystick'
 
 /** Ids of the furniture in the room currently being edited (the set the room
@@ -832,6 +833,7 @@ export default function App() {
         <TapeModeToggle />
         <Crosshair />
         <WalkJoystick />
+        <WalkCameraControls />
         <WalkHud />
         <DoorPrompt />
         <CatalogDrawer />